Change to predict/update terminology.
I originally used sense/update as my function names. But all the
literature uses update/predict, so I am normalizing on that.

I also wrote a test to ensure the update/predict functions in the
1d kalman filter chapter gives the same output as the full blown
kalman filter. Of course, they do. This is in the 1d_kf_compare
notebook. I will add this as an exercise in the multidimensional KF
chapter.