Altered PDF generation.
Added code to inject \appendix to the first appendix notebook
so that they are properly labelled. Started g-h chapter as chapter
one. There is probably a lot more I want to do now that I have
the basic idea of how to inject latex into the files before creating
the PDF, but this is a good start.