Chapter reordering; nonlinear chapter added.
I wrote a chapter introducing the problem of nonlinearity in KF because this
information is common to the EKF, UKF, etc. Then I reordered chapters. I want
people ot learn the UKF before the EKF. While I was doing this, I moved the
H Infinity and Ensemble filters to later chapters because they are mostly
curiosities in this book.