Fixed formatting for PDF.
I finally learned that the raw cells let you put latex directly
into the output. So now I have the appendix being created without
the [1] %%latex showing up in the output. Also wrote code to parse
the .tex file to make the Preface a chapter without numbering, so it
is now a proper preface, and not "chapter 0".

I have not touched the Windows generation tools, so PDF generation
in Windows is now broken. I don't think I care.