More info on mean, std, and variance.
I had information on these rather buried, and did not have a single example
of how to compute them. Now the first part of the chapter covers this much
more thoroughly.

Also added a section on plotting exponentials. Not sure I want to retain
it; it is a bit 'light'.