Added Bayesian information.
Major revamping - a lot of bayesian math has been added, and the
examples have been tweaked. I corrected a major confusion in the
update function - I was scaling, but said I was multipling by a
probability.

Also modifed the css for mathjax so that the font renders in a more
reasonable form.