Copy edit of UKF chapter.
Also found a mistake in the text about computing the residual
of angles that was duplicated in the EKF chapter, so I fixed the
error in that chapter as well.

I pulled out the 'track in a circle' example as it is the usual
sort of textbook nonsense problem that no one cares about. I created
an 'old-content' Notebook in case anyone still wants to see it; I
may put more content there as I continue to edit.