Issue #144 - update variable est/prediction names
I erroneously named the prediction of x x_est, which is confusing at best.

There were also issues with plotsize. The %matplotlib notebook magic
is resetting the size of the plots. Not sure if this has something to do
with some of my plotting code or not. Anyway, I wrote a context manager
to reset the plot size after exiting interactive mode.