Reformat, and changes to work with IPython 6.0
After a new Anaconda install I was getting a lot of minor errors
in the notebooks. I've fixed all of those. Along the ways I altered
the format of the PDF output to exclude the In[] Out[] tags so the
code is indented less. I also altered the font size of the notebooks
to better match fonts in other pages on chrome.

FilterPy was updated to 1.1.0, and this check in requires that release
at the minimum.