Formatting changes, and removal of some text
I am switching to the default look of Jupyter as much as possible.
Altering the look is not really supported by Jupyter, so rending
on GitHub (for example) is very different than when running in
your own browser.

I also got rid of the animations that really slow down execution
of the notebook in favor of using a slider to display a step in
a computation - the reader can see the plots 'animated' by moving
the slider. It is much faster to compute, and actually gives them
fine grained control to go forwards/backwards in time.

I am also revising my coverage of statistics, and moving the
details about Bayes and total conditional probability to the
third chapter. I shouldn't be talking about probability distributions
before covering that material.