Altering book format
Changed rest of chapters to use new formatting style. Also,
there is a bug in the current release of Jupyter that overwrites
the rcParams settings when you do the first import, so I had to
split the book format cell into two. This is actually better because
now the PDF, which strips out the formatting cell, will still have
the %matplotlib inline and import from future code.

Also, in the process of refactoring the information about Gaussian
and Bayes. Stuff is still strewn around too much. I need to make
this more streamlined. Chapters 2-4 talk about distributions, Bayes,
and Gaussians, and material is repeated a lot.