Remove redundant "dx += accel"
In Chapter 1, under "Exercise: The Effect of Acceleration", the solution utilizes the closed-form solution to the differential equations for acceleration, velocity, and displacement. Because this solution was used, there is no need to increment dx; in this solution, dx should just remain as its initial value (which we could call dx0, but I didn't change it here).